// frontend/src/app/api/models/pelada-resumo-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PeladaResumoResponse {
  cidade?: string;
  data?: string;
  estado?: string;
  horaFim?: string;
  horaInicio?: string;
  id?: number;
  localNome?: string;
  maxParticipantes?: number;
  nome?: string;
  organizadorNickname?: string;
  status?: 'AGENDADA' | 'CONFIRMADA' | 'EM_ANDAMENTO' | 'FINALIZADA' | 'CANCELADA';
  statusDescricao?: string;
  tipoCampo?: 'FUTSAL' | 'SOCIETY';
  tipoCampoDescricao?: string;
  totalConfirmados?: number;
  vagasRestantes?: number;
  valorPorJogador?: number;
}